backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
selenium==4.15.2
requests==2.31.0
python-multipart==0.0.6
fastapi
uvicorn[standard]
selenium
python-dotenv